Extract enterprise capability rule module
