+++ tests/unit/app/test_enterprise_readiness_additional.py
@@ -4,7 +4,7 @@ from datetime import datetime
 import pytest
 from fastapi import Response
 
-from app import enterprise_feature_flags, enterprise_runtime_config
+from app import enterprise_capability_rules, enterprise_feature_flags, enterprise_runtime_config
 from app.enterprise_readiness import (
     _ACTOR_ID_HEADER,
     _AUDIT_ACCESS_MODE_PRIVILEGED_READ,
@@ -210,6 +210,11 @@ def test_enterprise_readiness_reexports_runtime_config_boundary():
 
 def test_enterprise_readiness_reexports_feature_flag_boundary():
     assert _feature_flag_enabled is enterprise_feature_flags._feature_flag_enabled
+
+
+def test_enterprise_readiness_reexports_capability_rule_boundary():
+    assert _required_capability is enterprise_capability_rules._required_capability
+    assert load_capability_rules is enterprise_capability_rules.load_capability_rules
 
 
 def test_load_json_map_fails_closed_for_missing_invalid_or_non_object_json(monkeypatch):
